requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Weaviate
weaviate-client==4.9.3

# Machine Learning & NLP
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.46.3

# Vietnamese NLP
pyvi==0.1.1

# Ranking & Search
rank-bm25==0.2.2

# Google Gemini
google-generativeai==0.8.3

# Data Processing
numpy==1.26.4
tqdm==4.67.1

# Utilities
unicodedata2==15.1.0
